--- src/compile/mod.rs
@@ -1,3 +1,0 @@
-pub mod errors;
-mod output;
-pub mod project;
